Added create_main_module() and a helper method for template files
- added create_main_module() which renders the template __main__.py file
to the target directory
- added create_file_from_template() which takes paths to the template
and target directories and a filename and renders the template file to a
file of the same name in the target directory. It's basically a more
specific wrapper around render_template_to_file() that assumes the file
is going to be named the same thing when it's copied over. It reduces 4
commonly repeated lines to a single function call, so that's nice