Removed old TODOs
Removed TODOs about initialize.py input validation and output. If
there's an error thrown, then the expectation is that execution stops
and the error is printed. The package name validation should be enough,
and the user is specifically asked to use a valid format despite the
fact that invalid syntax will be corrected