Further development on BrowserStack support
- WebDriverTestCase._bs_driver_init() sets CAPABILITIES['name'] to the
test method name of the current test. This identifier should show up on
BrowserStack to help organize test results
- Extracted selenium package version number to version.__selenium__.
This will make it easier to ensure the version of selenium used with
BrowserStack matches the local version for consistent results
- Added BS_CAPABILITIES and set_browserstack_capabilities() to
BrowserStackConfig. Currently not yet implemented, but the goal is to
configure BrowserStack capabilities for projects in addition to
per-browser desired capabilities